+++ mobx_codegen/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.3.9+1 - 0.3.10
+
+- Alters the analyzer dependency to support a range — from the previously supported version
+  (0.36.3), up to latest (0.39.0).
+- Adds support for library prefixes in all situations (`import 'package:foo' as foo`),
+  so type names are prefixed in generated part files.
+
 ## 0.3.8 - 0.3.9+1
 
 - Fixes a minor issue where types in generated code would appear as dynamic when they shouldn't.